+++ App/Model/Auth/UserFactory.php
@@ -18,7 +18,7 @@ class UserFactory
     const KEY_USER_NAME = 'name';
     const KEY_USER_PASSWORD = 'password';
     const KEY_USER_EMAIL = 'email';
-    const KEY_USER_TOKEN = 'remember_token';
+    const KEY_USER_TOKEN = 'access_token';
     const KEY_SUPERUSER = 'superuser';
     const KEY_ROLES = 'roles';
     const TABLE_USER = 'users';
@@ -33,5 +33,6 @@ class UserFactory
     public static function getUser($user = [])
     {
         return new NormalUser($user);
+//        return new OAuthUser($user);  //oauth
     }
 }